+++ src/main/scala/com/decisionbrain/cplex/cp/Segment.scala
@@ -7,7 +7,7 @@
 package com.decisionbrain.cplex.cp
 
 /**
-  * This class represents a segment of a function.
+  * This class represents a segment of a piecewise linear function.
   *
   * @param start is the start of the segment
   * @param end is the end of the segment
@@ -15,29 +15,7 @@ package com.decisionbrain.cplex.cp
   * @param valueRight is the value of the function at the right-most point of the segment
   * @tparam T is the type of the value
   */
-class Segment[T](val start: T, val end: T, val valueLeft: T, val valueRight: T) {
-
-  /**
-    * Constructor of a segment for a step function
-    *
-    * @param start is the start of the segment
-    * @param end is the end of the segment
-    * @param value is the value of the function on the segment
-    */
-  def this(start: T, end: T, value: T) = {
-    this(start, end, value, value)
-  }
-
-  /**
-    * Returns the value of the step function on the segment. Throws an exceptoin if it is not a segment of a step
-    * function
-    *
-    * @return
-    */
-  def value = {
-    if (valueLeft != valueRight) throw new RuntimeException("Not a step function")
-    valueLeft
-  }
+class Segment[T <: AnyVal](val start: T, val end: T, val valueLeft: T, val valueRight: T) {
 
   /**
     * Converts the segment in a human readable format.
@@ -46,10 +24,28 @@ class Segment[T](val start: T, val end: T, val valueLeft: T, val valueRight: T) 
     */
   override def toString(): String = {
     val builder = new StringBuilder
-    if (valueLeft == valueRight)
-      builder.append("[").append(start).append(" .. ").append(end).append(") -> ").append(valueLeft)
-    else
-      builder.append("[").append(start).append(" .. ").append(end).append(") -> [").append(valueLeft).append(" .. ").append(valueRight).append(")")
+    builder.append("[").append(start).append(" .. ").append(end).append(") -> [").append(valueLeft).append(" .. ").append(valueRight).append(")")
+    builder.toString
+  }
+}
+
+/**
+  * This class represent a segment of a step function, a cumul function or a state function.
+  *
+  * @param start is the start of the segment
+  * @param end is the end of the segment
+  * @param value is the value of the function on the segment
+  * @tparam T is the type of the value
+  */
+class ConstantSegment[T <: AnyVal](start: T, end: T, val value: T) extends Segment[T](start, end, value, value) {
+  /**
+    * Converts the segment in a human readable format.
+    *
+    * @return a character string
+    */
+  override def toString(): String = {
+    val builder = new StringBuilder
+    builder.append("[").append(start).append(" .. ").append(end).append(") -> ").append(valueLeft)
     builder.toString
   }
 }